basic 1 complete

+++ lab5/include/priority_queue.h
@@ -1,8 +1,8 @@
 #pragma once
-int add_node(void (*callback_f)(),void* arguments,int times, int time_gap);
+unsigned long long add_node(void (*callback_f)(),void* arguments,unsigned long long times, unsigned long long time_gap);
 struct node* delete_first_node();
 struct node{
-    int time_to_ring;
+    unsigned long long time_to_ring;
     void (*todo)();
     void *arguments;
     struct node* next;
